requirements.txt
streamlit
pandas
numpy
plotly
openpyxl
google-generativeai
Pillow
pypdf
